requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
langchain-core==1.2.28
langchain-google-genai==4.2.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
google-generativeai==0.8.6
pymongo==4.15.5
motor==3.7.1
pinecone==8.1.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
structlog==25.5.0
requests==2.33.0
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
pytest-mock==3.15.1
langgraph-checkpoint-mongodb==0.3.1
slowapi==0.1.9
redis==7.4.0
limits==5.8.0
python-jose[cryptography]==3.5.0
types-python-jose
bcrypt==5.0.0
python-multipart==0.0.22
email-validator==2.3.0
ruff==0.3.5
pre-commit==3.7.0
